Moved DependOnPropertyAttribute to a new file
